--- FirstProject/src/components/MyInfo.jsx
@@ -1,11 +1,0 @@
-import React from "react";
-
-function MyInfo() {
-  return (
-    <div>
-      <h3>MyInfo</h3>
-    </div>
-  );
-}
-
-export default MyInfo;
