+++ FirstProject/src/routes/Movie.jsx
@@ -1,6 +1,6 @@
 import { BrowserRouter as Router, Switch, Route } from "react-router-dom";
-import Detail from "./routes/Detail";
-import Home from "./routes/Home";
+import Detail from "../movie/route/Detail";
+import Home from "../movie/route/Home";
 
 function Movie() {
   return (
