+++ FirstProject/src/routes/RegisterPage.jsx
@@ -1,46 +1,57 @@
-import React, { useState } from 'react';
+import React, {useState} from "react";
 import { useDispatch } from 'react-redux';
-import { loginUser } from '../../../_actions/user_action';
-import { useHistory } from 'react-router-dom';
+import { registerUser } from '../_actions/user_action';
+import { useHistory } from "react-router-dom";
 
-
-function LoginPage() {
+function RegisterPage(){
     const history=useHistory()
-
     const dispatch=useDispatch()
 
     const [Email, setEmail]=useState("")
+    const [Name, setName]=useState("")
     const [Password, setPassword]=useState("")
+    const [ConfirmPassword, setConfirmPassword]=useState("")
 
     const onEmailHandler=(e)=>{
         setEmail(e.currentTarget.value)
+    }
+
+    const onNameHandler=(e)=>{
+        setName(e.currentTarget.value)
     }
 
     const onPasswordHandler=(e)=>{
         setPassword(e.currentTarget.value)
     }
 
+    const onConfirmPasswordHandler=(e)=>{
+        setConfirmPassword(e.currentTarget.value)
+    }
+
     const onSubmitHandler=(e)=>{
         e.preventDefault();
 
+        if(Password !== ConfirmPassword){
+            return alert('비밀번호확인이 일치하지 않습니다.')
+        }
+
         let body={
             email: Email,
+            name: Name,
             password: Password
         }
 
-        dispatch(loginUser(body))
+        dispatch(registerUser(body))
             .then(response=>{
-                if(response.payload.loginSuccess){
-                    alert('로그인 성공')
-                    history.push('/')
+                if(response.payload.success){
+                    history.push("/login")
                 }else{
-                    alert('Error')
+                    alert('회원가입실패')
                 }
             })
         
     }
-
-    return (
+    return(
         <div style={{
             display: 'flex', justifyContent: 'center', alignItems: 'center',
             width: '300px', height: '500px'
@@ -50,14 +61,23 @@ function LoginPage() {
             >
                 <label>Email</label>
                 <input type="email" value={Email} onChange={onEmailHandler}/>
+
+                <label>Name</label>
+                <input type="text" value={Name} onChange={onNameHandler}/>
+
                 <label>Password</label>
                 <input type="password" value={Password} onChange={onPasswordHandler}/>
+
+                <label>Confirm Password</label>
+                <input type="password" value={ConfirmPassword} onChange={onConfirmPasswordHandler}/>
+
+
                 <button>
-                    Login
+                    Join
                 </button>
             </form>
         </div>
     )
 }
 
-export default LoginPage
+export default RegisterPage
